Add files via upload
Dashboards are a way to monitor your business and see all your most important metrics at a glance. The visualizations on a dashboard can come from one underlying semantic model or many, and from one underlying report or many. A dashboard combines on-premises and cloud data, providing a consolidated view of data.

A dashboard isn't just a pretty picture. It's highly interactive, and the tiles update as the underlying data changes.